+++ tests/perf/index.js
@@ -1,4 +1,4 @@
 
-require('./construction.js');
-require('./property_access.js');
-require('./method_calls.js');
+import './construction.js';
+import './property_access.js';
+import './method_calls.js';