+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 
 # Changelog
 
+- v2.0
+  - Upgrade `proxy-extend` to use modern Node.js 14+ features.
+  - Convert package to ES modules by default using `module: "type"`.
+  - Use `exports` in `package.json` rather than `main`. This is technically a breaking change due to the change
+    in which subpaths can be imported.
+  - Drop support for Node 12.
+  - Upgrade to package-lock v2 format.
+
 - v1.3.x
   - Drop support for Node v10, and IE 11.